Moved src in dt-fixer to conform to standard Maven layout.
