+++ dt-fixer/src/main/java/edu/washington/cs/dt/fixer/Main/MethodAndStaticStmt.java
